services/urban-events-graphql-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
graphene==3.3
starlette-graphene3==0.6.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2